requirements.txt
llama-parse
datasets
huggingface_hub
python-dotenv
